feat:One_lotto -> Empty_lotto 로 분리
