+++ .idea/workspace.xml
@@ -5,8 +5,9 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="75cfef9c-bc50-442b-99d5-2d9016fefebb" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/src/main/java/kr/codesquad/Calculator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kr/codesquad/Calculator.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/kr/codesquad/Prizes.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kr/codesquad/Prizes.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/kr/codesquad/Empty_lotto.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kr/codesquad/Empty_lotto.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/kr/codesquad/Jackpot_Handler.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kr/codesquad/Jackpot_Handler.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/kr/codesquad/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/kr/codesquad/Main.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -101,7 +102,7 @@
       <workItem from="1672210432725" duration="156000" />
       <workItem from="1672642153535" duration="4115000" />
       <workItem from="1672646622277" duration="6789000" />
-      <workItem from="1672662103062" duration="11283000" />
+      <workItem from="1672662103062" duration="11455000" />
     </task>
     <servers />
   </component>
